s-type/requirements.txt
pygame-ce
python-lsp-server
